samples/pdf_summary_agent/requirements.txt
google-adk
google-cloud-bigquery
pandas
db_dtypes
cloudpickle
dotenv